+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.12.0.pre.alpha (TBD)
+# 1.12.0 (TBD)
 
 Enhanced Job: